# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar TODOs",
        "title": "TODO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Muestra la página de bienvenida",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "general"
                ],
                "summary": "Bienvenida",
                "responses": {
                    "200": {
                        "description": "Página de bienvenida",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "description": "Obtiene la lista de todos los TODOs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Listar TODOs",
                "responses": {
                    "200": {
                        "description": "Lista de TODOs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todos.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo TODO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Crear TODO",
                "parameters": [
                    {
                        "description": "Datos del TODO",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "TODO creado",
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Obtiene un TODO por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Obtener TODO",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del TODO",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TODO encontrado",
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "TODO no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un TODO existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Actualizar TODO",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del TODO",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del TODO",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TODO actualizado",
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "TODO no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "todos.Todo": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}